# backend/src/Apps/Collections/openapi.yaml
openapi: 3.0.1
info:
  title: Dummy API (mojeEUC)
  description: Dummy service (mojeEUC) interface specification.
  termsOfService: https://cyberfox.cz/
  contact:
    name: Petr Tichý
    url: https://cyberfox.cz/
    email: petr.tichy@cyberfox.cz
  license:
    name: CyberFox Closed Source Commercial License (CFCSCL)
    url: https://cyberfox.cz/
  version: v1
servers:
  - url: https://cf-meuc-dev.azure-api.net/dummy/v1
paths:
...
  /book:
    post:
      tags:
        - Book
      summary: Creates a new entity.
      description: Creates a new entity.
      operationId: BookCreate
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
            example:
              authorId: string
              name: string
              originalPrice: 0
              publishYear: 0
      responses:
        '201':
          description: Book was created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
              example:
                authorId: string
                name: string
                originalPrice: 0
                publishYear: 0
                isActive: true
                id: string
                purchasedOn: string
        '400':
          description: Book creation request is invalid.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '410':
          description: "One of the following problems occurred: <br />\r\n            - Author does not exists or is not active <br />"
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    get:
      tags:
        - Book
      summary: Retrieves a list of books based on the specified criteria.
      description: Retrieves a list of books based on the specified criteria.
      operationId: BookList
      parameters:
        - name: onlyActive
          in: query
          description: Limits the resulting list to only active books.
          schema:
            type: boolean
            default: true
        - name: authorId
          in: query
          description: The ID of the author to filter the books by.
          schema:
            type: string
        - name: if-Modified-Since
          in: header
          description: Format - date-time (as date-time in RFC3339). Limits the resulting list to only books that have been modified since.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - string
        '304':
          description: Any book were modified since requested date.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
  '/book/{id}':
    delete:
      tags:
        - Book
      summary: Deletes entity by ID.
      description: Deletes entity by ID.
      operationId: BookDelete
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Book was deleted successfully.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Book was not found.
        '409':
          description: Book is not active.
        '412':
          description: Book was modified by another request.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    get:
      tags:
        - Book
      summary: Gets entity by ID.
      description: Gets entity by ID.
      operationId: BookGet
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: onlyActive
          in: query
          description: ''
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Author was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
              example:
                authorId: string
                name: string
                originalPrice: 0
                publishYear: 0
                isActive: true
                id: string
                purchasedOn: string
        '304':
          description: Book was not modified.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Book was not found.
        '409':
          description: Book is not active (if requested onlyActive record).
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    put:
      tags:
        - Book
      summary: Returns a list of entities.
      description: Returns a list of entities.
      operationId: BookUpdate
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
            example:
              authorId: string
              name: string
              originalPrice: 0
              publishYear: 0
      responses:
        '200':
          description: Book was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
              example:
                authorId: string
                name: string
                originalPrice: 0
                publishYear: 0
                isActive: true
                id: string
                purchasedOn: string
        '400':
          description: Book update request is invalid.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Book was not found.
        '409':
          description: Book is not active.
        '410':
          description: "One of the following problems occurred: <br />\r\n            - Author does not exists or is not active <br />"
        '412':
          description: Book was modified by another request.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
...
  /patron:
    post:
      tags:
        - Patron
      summary: Creates a new entity.
      description: Creates a new entity.
      operationId: PatronCreate
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatronCreateRequest'
            example:
              firstName: string
              lastName: string
              dateOfBirth: string
      responses:
        '201':
          description: Patron was created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatronResponse'
              example:
                firstName: string
                lastName: string
                dateOfBirth: string
                isActive: true
                id: string
        '400':
          description: Patron creation request is invalid.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    get:
      tags:
        - Patron
      summary: Lists all patrons.
      description: Lists all patrons.
      operationId: PatronList
      parameters:
        - name: onlyActive
          in: query
          description: ''
          schema:
            type: boolean
            default: true
        - name: query
          in: query
          description: ''
          schema:
            type: string
        - name: if-Modified-Since
          in: header
          description: Format - date-time (as date-time in RFC3339).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - string
        '304':
          description: Any patron were modified since requested date.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
  '/patron/{id}':
    delete:
      tags:
        - Patron
      summary: Deletes entity by ID.
      description: Deletes entity by ID.
      operationId: PatronDelete
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Patron was deleted successfully.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Patron was not found.
        '409':
          description: Patron is not active.
        '412':
          description: Patron was modified by another request.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    get:
      tags:
        - Patron
      summary: Gets entity by ID.
      description: Gets entity by ID.
      operationId: PatronGet
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: onlyActive
          in: query
          description: ''
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Patron was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatronResponse'
              example:
                firstName: string
                lastName: string
                dateOfBirth: string
                isActive: true
                id: string
        '304':
          description: Patron was not modified.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Patron was not found.
        '409':
          description: Patron is not active.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    put:
      tags:
        - Patron
      summary: Returns a list of entities.
      description: Returns a list of entities.
      operationId: PatronUpdate
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatronUpdateRequest'
            example:
              firstName: string
              lastName: string
              dateOfBirth: string
      responses:
        '200':
          description: Patron was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatronResponse'
              example:
                firstName: string
                lastName: string
                dateOfBirth: string
                isActive: true
                id: string
        '400':
          description: Patron update request is invalid.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Patron was not found.
        '409':
          description: Patron is not active.
        '412':
          description: Patron was modified by another request.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
...
components:
  schemas:
...
    BookCreateRequest:
      required:
        - authorId
        - name
        - originalPrice
        - publishYear
      type: object
      properties:
        authorId:
          minLength: 1
          type: string
          description: Tha author of the book
        name:
          minLength: 1
          type: string
          description: The Name of the book
        originalPrice:
          type: number
          description: The original price of the book.
          format: double
        publishYear:
          type: integer
          description: The year the book was published.
          format: int32
      additionalProperties: false
      description: Request to create a book.
    BookResponse:
      required:
        - authorId
        - id
        - isActive
        - name
        - originalPrice
        - publishYear
        - purchasedOn
      type: object
      properties:
        authorId:
          minLength: 1
          type: string
          description: Tha author of the book
        name:
          minLength: 1
          type: string
          description: The Name of the book
        originalPrice:
          type: number
          description: The original price of the book.
          format: double
        publishYear:
          type: integer
          description: The year the book was published.
          format: int32
        isActive:
          type: boolean
          description: Indicates whether the book is active (or soft-deleted).
        id:
          minLength: 1
          type: string
          description: Book ID.
        purchasedOn:
          type: string
          description: The date the book was purchased.
          format: date-time
      additionalProperties: false
      description: General response to book requests.
    BookUpdateRequest:
      required:
        - authorId
        - name
        - originalPrice
        - publishYear
      type: object
      properties:
        authorId:
          minLength: 1
          type: string
          description: Tha author of the book
        name:
          minLength: 1
          type: string
          description: The Name of the book
        originalPrice:
          type: number
          description: The original price of the book.
          format: double
        publishYear:
          type: integer
          description: The year the book was published.
          format: int32
      additionalProperties: false
      description: Request to update a book.
...
    PatronCreateRequest:
      required:
        - dateOfBirth
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          minLength: 1
          type: string
          description: The first name of the patron.
        lastName:
          minLength: 1
          type: string
          description: The last name of the patron.
        dateOfBirth:
          type: string
          description: Date of birth of the patron.
          format: date-time
      additionalProperties: false
      description: Request to create a patron.
    PatronResponse:
      required:
        - dateOfBirth
        - firstName
        - id
        - isActive
        - lastName
      type: object
      properties:
        firstName:
          minLength: 1
          type: string
          description: The first name of the patron.
        lastName:
          minLength: 1
          type: string
          description: The last name of the patron.
        dateOfBirth:
          type: string
          description: Date of birth of the patron.
          format: date-time
        isActive:
          type: boolean
          description: Indicates whether the patron is active (or soft-deleted).
        id:
          minLength: 1
          type: string
          description: Patron ID.
      additionalProperties: false
      description: General response to patron requests.
    PatronUpdateRequest:
      required:
        - dateOfBirth
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          minLength: 1
          type: string
          description: The first name of the patron.
        lastName:
          minLength: 1
          type: string
          description: The last name of the patron.
        dateOfBirth:
          type: string
          description: Date of birth of the patron.
          format: date-time
      additionalProperties: false
      description: Request to update a patron.
...
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      name: Ocp-Apim-Subscription-Key
      in: header
    apiKeyQuery:
      type: apiKey
      name: subscription-key
      in: query
security:
  - apiKeyHeader: [ ]
  - apiKeyQuery: [ ]